+++ src/configure/resources/telemetryKeys.ts
@@ -28,10 +28,13 @@ export class TelemetryKeys {
     public static SelectedTemplateType = 'selectedTemplateType';
     public static WorkflowFileName = 'workflowFileName';
     public static GitHubRepoCreated = 'GitHubRepoCreated';
-
+    public static IsErrorWhitelisted = 'IsErrorWhitelisted';
+    public static FF_UseGithubForCreatingNewRepository = 'FF_UseGithubForCreatingNewRepository';
+    public static FF_UseAzurePipelinesForGithub = 'FF_UseAzurePipelinesForGithub';
 
     public static resourceType = 'resourceType';
     public static resourceKind = 'resourceKind';
+    public static resourceIdHash = 'resourceIdHash';
 
     // Durations
     public static ExtensionActivationDuration = 'extensionActivationDuration';